# PERFORMANCE: Improve speed of saving when there are lots of scripts